COG-1107: UserService interface is extracted
